+++ help/csharp/HowToHandshakeUtils/parameters/ImageParameter.cs
@@ -7,7 +7,7 @@ namespace RCP.Parameters
     {
         public new ImageDefinition TypeDefinition => base.TypeDefinition as ImageDefinition;
 
-        public ImageParameter(Int16 id, IParameterManager manager, ImageDefinition typeDefinition) 
+        public ImageParameter(int id, IParameterManager manager, ImageDefinition typeDefinition) 
             : base(id, manager, typeDefinition)
         {
         }
