+++ help/csharp/HowToHandshakeUtils/typedefinitions/BooleanDefinition.cs
@@ -15,7 +15,7 @@ namespace RCP.Types
         {
         }
 
-        public override Parameter CreateParameter(short id, IParameterManager manager) => new ValueParameter<bool>(id, manager, this);
+        public override Parameter CreateParameter(int id, IParameterManager manager) => new ValueParameter<bool>(id, manager, this);
 
         public override bool ReadValue(KaitaiStream input)
         {
