+++ help/csharp/HowToHandshakeUtils/typedefinitions/EnumDefinition.cs
@@ -39,7 +39,7 @@ namespace RCP.Types
             }
         }
 
-        public override Parameter CreateParameter(short id, IParameterManager manager) => new EnumParameter(id, manager, this);
+        public override Parameter CreateParameter(int id, IParameterManager manager) => new EnumParameter(id, manager, this);
 
         public override void ResetForInitialize()
         {
